chore: refactor routes
