+++ src/server/routes/search.ts
@@ -1,243 +1,47 @@
-import { Hono, type Context } from "hono";
+import { Hono } from "hono";
 import {
-  getActiveWebEngines,
   getCustomEngineTypes,
-  getEngineRegistry,
   getEnginesForCustomType,
-  getEnginesForSearchType as getEnginesForType,
 } from "../extensions/engines/registry";
 import {
   getSearchResultTabById,
   getSearchResultTabs,
 } from "../extensions/search-result-tabs/registry";
-import { getSlotPlugins } from "../extensions/slots/registry";
 import {
   createSearchEngineContext,
-  fetchRelatedSearches,
   mergeNewResults,
-  scoreResults,
   search,
   searchSingleEngine,
 } from "../search";
 import {
-  SLOT_POSITION_SETTING_KEY,
-  SlotPanelPosition,
-  type EngineConfig,
   type EngineTiming,
   type ScoredResult,
-  type SearchResponse,
   type SearchType,
-  type SlotPanelResult,
-  type SlotPluginContext,
   type TimeFilter,
 } from "../types";
 import * as cache from "../utils/cache";
-import { getLocale } from "../utils/hono";
 import { logger } from "../utils/logger";
-import { outgoingFetch } from "../utils/outgoing";
-import { asString, getSettings, isDisabled } from "../utils/plugin-settings";
-import { checkRateLimit } from "../utils/rate-limit";
+import { isDisabled } from "../utils/plugin-settings";
 import { getClientIp } from "../utils/request";
-import { injectScope, translateHTML } from "../utils/translation";
+import {
+  _applyRateLimit,
+  cacheKey,
+  isValidQuery,
+  parseEngineConfig,
+} from "../utils/search";
 
-const DEGOOG_SETTINGS_ID = "degoog-settings";
 const router = new Hono();
-
-const _applyRateLimit = async (c: Context): Promise<Response | null> => {
-  const settings = await getSettings(DEGOOG_SETTINGS_ID);
-  const opts: Record<string, string> = {};
-  for (const [k, v] of Object.entries(settings)) {
-    opts[k] = typeof v === "string" ? v : Array.isArray(v) ? (v[0] ?? "") : "";
-  }
-  if (opts.rateLimitEnabled !== "true") return null;
-  const ip = getClientIp(c) ?? "unknown";
-  const result = checkRateLimit(ip, opts);
-  if (!result.allowed && result.retryAfterSec !== undefined) {
-    return c.json({ error: "Too many requests" }, 429, {
-      "Retry-After": String(result.retryAfterSec),
-    });
-  }
-  return null;
-};
-
-function parseEngineConfig(query: URLSearchParams): EngineConfig {
-  const registry = getEngineRegistry();
-  const config: EngineConfig = {};
-  for (const { id } of registry) {
-    config[id] = query.get(id) !== "false";
-  }
-  return config;
-}
-
-const DEFAULT_LANGUAGES = [
-  "af",
-  "am",
-  "ar",
-  "az",
-  "be",
-  "bg",
-  "bn",
-  "bs",
-  "ca",
-  "cs",
-  "cy",
-  "da",
-  "de",
-  "el",
-  "en",
-  "eo",
-  "es",
-  "et",
-  "eu",
-  "fa",
-  "fi",
-  "fr",
-  "ga",
-  "gl",
-  "gu",
-  "he",
-  "hi",
-  "hr",
-  "hu",
-  "hy",
-  "id",
-  "is",
-  "it",
-  "ja",
-  "ka",
-  "kk",
-  "km",
-  "kn",
-  "ko",
-  "ku",
-  "ky",
-  "lb",
-  "lo",
-  "lt",
-  "lv",
-  "mk",
-  "ml",
-  "mn",
-  "mr",
-  "ms",
-  "my",
-  "ne",
-  "nl",
-  "no",
-  "or",
-  "pa",
-  "pl",
-  "ps",
-  "pt",
-  "ro",
-  "ru",
-  "sd",
-  "si",
-  "sk",
-  "sl",
-  "so",
-  "sq",
-  "sr",
-  "st",
-  "sv",
-  "sw",
-  "ta",
-  "te",
-  "tg",
-  "th",
-  "tk",
-  "tl",
-  "tr",
-  "uk",
-  "ur",
-  "uz",
-  "vi",
-  "xh",
-  "yi",
-  "yo",
-  "zh",
-  "zu",
-];
-
-function cacheKey(
-  query: string,
-  engines: EngineConfig,
-  type: SearchType,
-  page: number,
-  timeFilter: TimeFilter = "any",
-  lang = "",
-  dateFrom = "",
-  dateTo = "",
-): string {
-  const q = query.trim().toLowerCase();
-  return `${q}|${JSON.stringify(engines)}|${type}|${page}|${timeFilter}|${lang}|${dateFrom}|${dateTo}`;
-}
-
-async function runSlotPlugins(
-  query: string,
-  clientIp?: string,
-  results?: ScoredResult[],
-  options?: { excludePosition?: SlotPanelPosition; locale?: string },
-): Promise<SlotPanelResult[]> {
-  const plugins = getSlotPlugins();
-  const panels: SlotPanelResult[] = [];
-  const exclude = options?.excludePosition;
-  const locale = options?.locale;
-  for (const plugin of plugins) {
-    const slotSettingsId = plugin.settingsId ?? `slot-${plugin.id}`;
-    let effectivePosition: SlotPanelPosition = plugin.position;
-    if (plugin.slotPositions?.length) {
-      const raw = await getSettings(slotSettingsId);
-      const chosen = asString(raw[SLOT_POSITION_SETTING_KEY]);
-      if (
-        chosen &&
-        plugin.slotPositions.includes(chosen as SlotPanelPosition)
-      ) {
-        effectivePosition = chosen as SlotPanelPosition;
-      }
-    }
-    if (exclude && effectivePosition === exclude) continue;
-    try {
-      if (await isDisabled(slotSettingsId)) continue;
-      const ok = await Promise.resolve(plugin.trigger(query.trim()));
-      if (!ok) continue;
-      if (plugin.t && locale) plugin.t.setLocale(locale);
-      const context: SlotPluginContext = {
-        clientIp,
-        results: plugin.waitForResults ? results : undefined,
-        fetch: outgoingFetch as SlotPluginContext["fetch"],
-      };
-      const t0 = performance.now();
-      const out = await plugin.execute(query, context);
-      logger.debug(
-        "plugin",
-        `${plugin.id} executed in ${Math.round(performance.now() - t0)}ms`,
-      );
-      if (!out.html || !out.html.trim()) continue;
-      panels.push({
-        id: plugin.id,
-        title: out.title,
-        html: injectScope(
-          plugin.t ? translateHTML(out.html, plugin.t) : out.html,
-          `slots/${plugin.id}`,
-        ),
-        position: effectivePosition,
-        gridSize: plugin.gridSize,
-      });
-    } catch {}
-  }
-  return panels;
-}
 
 router.get("/api/search", async (c) => {
   const limitRes = await _applyRateLimit(c);
   if (limitRes) return limitRes;
-  const searchType = (c.req.query("type") || "web") as SearchType;
-  let query = c.req.query("q") ?? "";
-  if (typeof query !== "string") query = "";
-  if (!query.trim())
-    return c.json({ error: "Missing query parameter 'q'" }, 400);
 
+  const query = c.req.query("q") ?? "";
+
+  if (!isValidQuery(query))
+    return c.json({ error: "Missing or invalid query parameter 'q'" }, 400);
+
+  const searchType = (c.req.query("type") || "web") as SearchType;
   const engines = parseEngineConfig(new URL(c.req.url).searchParams);
   const page = Math.max(
     1,
@@ -259,341 +63,27 @@ router.get("/api/search", async (c) => {
   );
 
   const cached = cache.get(key);
-  let response: SearchResponse;
-  if (cached) {
-    response = cached;
-  } else {
-    response = await search(
-      query,
-      engines,
-      searchType,
-      page,
-      timeFilter,
-      lang,
-      dateFrom,
-      dateTo,
-    );
-    const ttl = cache.hasFailedEngines(response)
-      ? cache.SHORT_TTL_MS
-      : searchType === "news"
-        ? cache.NEWS_TTL_MS
-        : undefined;
-    cache.set(key, response, ttl);
-  }
+  if (cached) return c.json(cached);
+
+  const response = await search(
+    query,
+    engines,
+    searchType,
+    page,
+    timeFilter,
+    lang,
+    dateFrom,
+    dateTo,
+  );
+
+  const ttl = cache.hasFailedEngines(response)
+    ? cache.SHORT_TTL_MS
+    : searchType === "news"
+      ? cache.NEWS_TTL_MS
+      : undefined;
+  cache.set(key, response, ttl);
 
   return c.json(response);
-});
-
-router.get("/api/search/stream", async (c) => {
-  const limitRes = await _applyRateLimit(c);
-  if (limitRes) return limitRes;
-  const searchType = (c.req.query("type") || "web") as SearchType;
-  let query = c.req.query("q") ?? "";
-  if (typeof query !== "string") query = "";
-  if (!query.trim())
-    return c.json({ error: "Missing query parameter 'q'" }, 400);
-
-  const engines = parseEngineConfig(new URL(c.req.url).searchParams);
-  const page = Math.max(
-    1,
-    Math.min(10, Math.floor(Number(c.req.query("page"))) || 1),
-  );
-  const timeFilter = (c.req.query("time") || "any") as TimeFilter;
-  const lang = c.req.query("lang") || "";
-  const dateFrom = c.req.query("dateFrom") || "";
-  const dateTo = c.req.query("dateTo") || "";
-  const key = cacheKey(
-    query,
-    engines,
-    searchType,
-    page,
-    timeFilter,
-    lang,
-    dateFrom,
-    dateTo,
-  );
-
-  const cached = cache.get(key);
-  if (cached) {
-    const encoder = new TextEncoder();
-    const body = new ReadableStream({
-      start(controller) {
-        for (const et of cached.engineTimings) {
-          controller.enqueue(
-            encoder.encode(
-              `event: engine-result\ndata: ${JSON.stringify({
-                engine: et.name,
-                timing: et,
-                results: cached.results,
-                retry: false,
-                attempt: 0,
-              })}\n\n`,
-            ),
-          );
-        }
-        controller.enqueue(
-          encoder.encode(
-            `event: done\ndata: ${JSON.stringify({
-              totalTime: cached.totalTime,
-              engineTimings: cached.engineTimings,
-              relatedSearches: cached.relatedSearches,
-            })}\n\n`,
-          ),
-        );
-        controller.close();
-      },
-    });
-    return new Response(body, {
-      headers: {
-        "Content-Type": "text/event-stream",
-        "Cache-Control": "no-cache",
-        Connection: "keep-alive",
-      },
-    });
-  }
-
-  const settings = await getSettings(DEGOOG_SETTINGS_ID);
-  const autoRetry = asString(settings.streamingAutoRetry) === "true";
-  const maxRetries = Math.min(
-    5,
-    Math.max(1, parseInt(asString(settings.streamingMaxRetries) || "2", 10)),
-  );
-
-  const builtinTypes = new Set(["web", "images", "videos", "news"]);
-  const rawActiveEngines =
-    searchType === "web"
-      ? await getActiveWebEngines(engines)
-      : builtinTypes.has(searchType)
-        ? getEnginesForType(searchType, engines).map((e) => ({
-            id: e.id,
-            instance: e.instance,
-            score: 1,
-          }))
-        : getEnginesForCustomType(searchType).map((e) => ({
-            id: e.id,
-            instance: e.instance,
-            score: 1,
-          }));
-
-  if (rawActiveEngines.length === 0) {
-    return c.json({
-      results: [],
-      query,
-      totalTime: 0,
-      type: searchType,
-      engineTimings: [],
-      relatedSearches: [],
-    });
-  }
-
-  const start = performance.now();
-
-  let closed = false;
-
-  const stream = new ReadableStream({
-    start(controller) {
-      const encoder = new TextEncoder();
-      const allTimings: EngineTiming[] = [];
-      const allRawResults: {
-        results: import("../types").SearchResult[];
-        multiplier: number;
-      }[] = [];
-
-      function _send(event: string, data: unknown) {
-        if (closed) return;
-        try {
-          controller.enqueue(
-            encoder.encode(
-              `event: ${event}\ndata: ${JSON.stringify(data)}\n\n`,
-            ),
-          );
-        } catch {
-          closed = true;
-        }
-      }
-
-      const enginePromises = rawActiveEngines.map(
-        async ({ instance, score, id }) => {
-          const engineName = instance.name;
-          let attempt = 0;
-          let lastTiming: EngineTiming = {
-            name: engineName,
-            time: 0,
-            resultCount: 0,
-          };
-
-          while (attempt <= (autoRetry ? maxRetries : 0)) {
-            const isRetry = attempt > 0;
-            const { results, timing } = await searchSingleEngine(
-              id,
-              query,
-              page,
-              timeFilter,
-              lang,
-              dateFrom,
-              dateTo,
-            );
-            lastTiming = timing;
-
-            if (timing.resultCount > 0) {
-              allRawResults.push({ results, multiplier: score });
-              allTimings.push(timing);
-              _send("engine-result", {
-                engine: engineName,
-                timing,
-                results: scoreResults(allRawResults),
-                retry: isRetry,
-                attempt,
-              });
-              return;
-            }
-
-            attempt++;
-            if (attempt <= (autoRetry ? maxRetries : 0)) {
-              _send("engine-retry", {
-                engine: engineName,
-                attempt,
-                maxRetries,
-                timing,
-              });
-            }
-          }
-
-          allTimings.push(lastTiming);
-          _send("engine-result", {
-            engine: engineName,
-            timing: lastTiming,
-            results: scoreResults(allRawResults),
-            retry: false,
-            attempt: 0,
-          });
-        },
-      );
-
-      void Promise.all(enginePromises).then(async () => {
-        const totalTime = Math.round(performance.now() - start);
-        const finalResults = scoreResults(allRawResults);
-        let relatedSearches: string[] = [];
-        if (searchType === "web" && page === 1) {
-          relatedSearches = await fetchRelatedSearches(query).catch(
-            () => [] as string[],
-          );
-        }
-
-        const response: SearchResponse = {
-          results: finalResults,
-          query,
-          totalTime,
-          type: searchType,
-          engineTimings: allTimings,
-          relatedSearches,
-        };
-
-        const ttl = cache.hasFailedEngines(response)
-          ? cache.SHORT_TTL_MS
-          : searchType === "news"
-            ? cache.NEWS_TTL_MS
-            : undefined;
-        cache.set(key, response, ttl);
-
-        _send("done", {
-          totalTime,
-          engineTimings: allTimings,
-          relatedSearches,
-        });
-        if (!closed) {
-          closed = true;
-          controller.close();
-        }
-      });
-    },
-    cancel() {
-      closed = true;
-    },
-  });
-
-  return new Response(stream, {
-    headers: {
-      "Content-Type": "text/event-stream",
-      "Cache-Control": "no-cache",
-      Connection: "keep-alive",
-    },
-  });
-});
-
-router.post("/api/slots", async (c) => {
-  const limitRes = await _applyRateLimit(c);
-  if (limitRes) return limitRes;
-  let body: { query?: string; results?: ScoredResult[] };
-  try {
-    body = await c.req.json();
-  } catch {
-    return c.json({ error: "Invalid JSON" }, 400);
-  }
-  if (!body.query || !body.query.trim()) return c.json({ panels: [] });
-  const clientIp = getClientIp(c);
-  const panels = await runSlotPlugins(
-    body.query.trim(),
-    clientIp,
-    body.results,
-    {
-      excludePosition: SlotPanelPosition.AtAGlance,
-      locale: getLocale(c),
-    },
-  );
-  return c.json({ panels });
-});
-
-router.post("/api/slots/glance", async (c) => {
-  const limitRes = await _applyRateLimit(c);
-  if (limitRes) return limitRes;
-  let body: { query?: string; results?: ScoredResult[] };
-  try {
-    body = await c.req.json();
-  } catch {
-    return c.json({ error: "Invalid JSON" }, 400);
-  }
-  if (!body.query || !Array.isArray(body.results)) {
-    return c.json({ error: "Missing query or results" }, 400);
-  }
-  const clientIp = getClientIp(c);
-  const locale = getLocale(c);
-  const glancePlugins = getSlotPlugins().filter(
-    (p) => p.position === SlotPanelPosition.AtAGlance,
-  );
-  const panels: SlotPanelResult[] = [];
-  for (const plugin of glancePlugins) {
-    try {
-      const slotSettingsId = plugin.settingsId ?? `slot-${plugin.id}`;
-      if (await isDisabled(slotSettingsId)) continue;
-      const ok = await Promise.resolve(plugin.trigger(body.query!.trim()));
-      if (!ok) continue;
-      if (plugin.t && locale) plugin.t.setLocale(locale);
-      const context: SlotPluginContext = {
-        clientIp: clientIp ?? undefined,
-        results: body.results,
-        fetch: outgoingFetch as SlotPluginContext["fetch"],
-      };
-      const t0 = performance.now();
-      const out = await plugin.execute(body.query!.trim(), context);
-      logger.debug(
-        "plugin",
-        `${plugin.id} executed in ${Math.round(performance.now() - t0)}ms`,
-      );
-      if (!out.html || !out.html.trim()) continue;
-      panels.push({
-        id: plugin.id,
-        title: out.title,
-        html: injectScope(
-          plugin.t ? translateHTML(out.html, plugin.t) : out.html,
-          `slots/${plugin.id}`,
-        ),
-        position: plugin.position,
-        gridSize: plugin.gridSize,
-      });
-    } catch {}
-  }
-  return c.json({ panels });
 });
 
 router.get("/api/search/retry", async (c) => {
@@ -860,28 +350,6 @@ router.get("/api/tab-search", async (c) => {
       500,
     );
   }
-});
-
-router.get("/api/settings/languages", async (c) => {
-  const settings = await getSettings(DEGOOG_SETTINGS_ID);
-  if (asString(settings["languagesEnabled"] ?? "") !== "true") {
-    return c.json({ languages: DEFAULT_LANGUAGES });
-  }
-  const raw = asString(settings["languages"] ?? "");
-  const codes = raw
-    .split(/[\n,]/)
-    .map((s) => s.trim().toLowerCase())
-    .filter((s) => /^[a-z]{2,3}$/.test(s));
-  return c.json({ languages: codes.length > 0 ? codes : DEFAULT_LANGUAGES });
-});
-
-router.get("/api/settings/streaming", async (c) => {
-  const settings = await getSettings(DEGOOG_SETTINGS_ID);
-  return c.json({
-    enabled: asString(settings.streamingEnabled) === "true",
-    autoRetry: asString(settings.streamingAutoRetry) === "true",
-    maxRetries: parseInt(asString(settings.streamingMaxRetries) || "2", 10),
-  });
 });
 
 export default router;
